Resolve issue where result text wasn't removing from resultDiv
The issue was that I was removing the div container but it still had a
text node inside it so I was just adding onto the stack. realized I
could use .childNodes, an array of nodes, to pop the text node off if
resultDiv already existed.

Probably a little sloppy in the grand scheme but its working.